moving out minecraft_specs to droidlet for now
