+++ droidlet/dialog/craftassist/dialogue_objects/block_helpers.py
@@ -4,7 +4,7 @@ Copyright (c) Facebook, Inc. and its affiliates.
 import random
 import Levenshtein
 from droidlet.lowlevel.minecraft import block_data
-from droidlet.lowlevel.minecraft import specs as minecraft_specs
+from droidlet import minecraft_specs
 from droidlet.lowlevel.minecraft.mc_util import IDM
 
 # TODO FILTERS!
